+++ internlm/utils/evaluation.py
@@ -67,6 +67,7 @@ def evaluate_on_val_dls(
     logger,
     step_count,
     update_panel: bool = False,
+    streaming: bool = False,
 ):
     with switch_sequence_parallel_mode():
         torch.cuda.empty_cache()
@@ -75,7 +76,7 @@ def evaluate_on_val_dls(
         data_cfg = gpc.config.data
 
         for val_name, val_dl in val_dls.items():
-            if len(val_dl) == 0 and verbose:
+            if len(val_dl) == 0 and verbose and not streaming:
                 logger.info(f"Validation dataset: {val_name} is empty")
                 continue
 
@@ -91,7 +92,7 @@ def evaluate_on_val_dls(
             for val_idx, batch in tqdm(
                 enumerate(val_dl),
                 desc="Val.",
-                total=len(val_dl),
+                total=len(val_dl) if not streaming else None,
                 position=1,
                 disable=not verbose,
                 leave=False,
